refactor!: Fix URL scan & remove file scans
