+++ src/xiwen/config.py
@@ -1,15 +1,17 @@
 import os
 
+ASSETS_DIR = os.path.join(os.getcwd(), "src", "xiwen", "assets")
+# Test case (simplified hanzi)
+DEMO1 = "https://www.xuan-zang.com/bjzd"
+# Test case (traditional hanzi)
+DEMO2 = "https://www.xuan-zang.com/ttc"
+
 ENCODING = "utf-8"
 # Encoding used to force .csv files to adopt utf-8 from Pandas DataFrame
 ENCODING_HANZI = "utf_8_sig"
 HSK_GRADES = 7
 
-ASSETS_DIR = os.path.join(os.getcwd(), "assets")
-# Test case (simplified hanzi)
-BJZD = os.path.join(ASSETS_DIR, "beijingzhedie.txt")
-# Test case (traditional hanzi)
-TTC = os.path.join(ASSETS_DIR, "taoteching.txt")
+LOGS_DIR = os.path.join(os.getcwd(), "logs")
 
 RAW_DATA = os.path.join(os.getcwd(), "file_prep", "assets")
 HSK_PATH = os.path.join(RAW_DATA, "hsk30-chars-ext.csv")
@@ -35,10 +37,26 @@ Select option:
 
 DEMO_MESSAGE = """
 -> '10+' under 'HSK Grade' refers to hanzi beyond the HSK7-9 band.
--> 'Unique' columns capture the no. unique hanzi in the text per grade.
--> 'Count' columns capture the total no. hanzi per grade, duplicates included
-    ('今天天氣很好' = 5 unique hanzi, 6 total hanzi).")
--> '% of Total' gives the % of the left-hand value relative to all hanzi in the text.
--> 'Cumul No.' columns give the running totals per grade")
-    (1st 'Cumul. No.' column, HSK3 row gives the sum of unique characters found that belong to HSK1, HSK2, and HSK3).
+-> 'Unique' cols capture no. unique hanzi found per grade.
+-> 'Count' cols capture total no. hanzi found per grade.
+    '今天天氣很好' = 5 unique hanzi, 6 total hanzi.
+-> '% of Total' gives the % relative to all hanzi found.
+-> 'Cumul' cols give the cumulative totals per grade
+   'Cumul. Unique' col, HSK3 row gives the sum of unique hanzi found for HSK1, HSK2 and HSK3.
+"""
+
+EXPORT_OPTIONS = """
+"Select an option to export to .csv:
+-> 'a' = export all detected HSK hanzi (excludes outliers)
+-> 'c' = export custom HSK hanzi selection
+-> 'f' = export full HSK hanzi list
+-> 'o' = export detected outliers (non-HSK hanzi)
+-> 's' = export stats
+-> 'x' = exit to main screen
+"""
+
+CUSTOM_EXPORT = """
+Custom export:
+Enter the HSK grade(s) to export in any order
+-> e.g. to export HSK2 and HSK5 enter '25' or '52'
 """
