+++ src/xiwen/main.py
@@ -1,6 +1,5 @@
-import tkinter as tk
-from .config import WELCOME_MESSAGE, MENU_OPTIONS
-from utils.commands import cmd_demo, cmd_scan, cmd_url
+from config import WELCOME_MESSAGE, MENU_OPTIONS
+from utils.commands import cmd_demo, cmd_url
 
 
 def xiwen():
@@ -14,7 +13,7 @@ def xiwen():
         print(MENU_OPTIONS)
         command = input().upper()
 
-        if command not in ["D", "S", "U", "Q"]:
+        if command not in ["D", "U", "Q"]:
             continue  # Invalid command - repeat options
 
         elif command == "Q":  # Quit command
@@ -24,28 +23,10 @@ def xiwen():
             cmd_demo()
             continue
 
-        elif command == "S":  # Scan command
-            cmd_scan()
-            continue
-
         elif command == "U":  # URL command
             cmd_url()
             continue
 
 
-# Initialise Tkinter
-ROOT = tk.Tk()
-
-
-def handle_quit(self):
-    """Closes Tkinter window on quit"""
-    self.destroy()
-    exit()
-
-
 # Run program
 xiwen()
-
-# Exit Tkinter
-handle_quit(ROOT)
-ROOT.mainloop()
