+++ src/xiwen/utils/data.py
@@ -1,46 +1,17 @@
 import pandas as pd
-from pypdf import PdfReader
 from utils.hanzi import filter_text, partition_hanzi, identify, get_stats
-from xiwen.config import ENCODING
 
 
-def _extract_from_pdf(pdf_path):
+def process_data(html: str, hsk_simp: list, hsk_trad: list) -> tuple[list]:
     """
-    Extracts text from a PDF
-
-    Parameters
-    ----------
-    pdf_path : str
-        PDF filepath
-
-    Returns
-    -------
-    text : str
-        text extracted from PDF
-    """
-    text = ""
-
-    with open(pdf_path, "rb") as f:
-        reader = PdfReader(f)
-        for page in reader.pages:
-            text += page.extract_text()
-
-    return text
-
-
-def process_data(
-    target: str, hsk_simp: list, hsk_trad: list, html: bool = False
-) -> tuple[list]:
-    """
-    Searches for and extracts Chinese characters (hanzi) from a text file
+    Searches for and extracts Chinese characters (hanzi) from html
     according to whether they are simplified or traditional characters
     within the HSK1 to HSK6 vocabulary range, or other characters of either variant
 
     Parameters
     ----------
-    target : str
-        if html=False: path to file on device
-        if html=True: HTML extracted from URL
+    html : str
+        HTML extracted from URL
 
     hsk_simp : list
         all simplified hanzi from HSK1 to HSK9
@@ -48,12 +19,8 @@ def process_data(
     hsk_trad : list
         all traditional hanzi equivalents to hsk_simp
 
-    html : bool
-        flag True if html else False
-
     Returns
     -------  NOTE: all lists below include duplicates
-
     hanzi_list : list
         full list of hanzi found
 
@@ -69,20 +36,8 @@ def process_data(
     outliers : list
         all hanzi found that don't belong to the above lists
     """
-    if html:  # url=True, target is HTML
-        text = target
-
-    elif not html:
-        # Send PDFs to _extract_from_pdf
-        if target[-3:] == "pdf":
-            text = _extract_from_pdf(target)
-        # Directly extract from other file types
-        else:
-            with open(target, "r", encoding=ENCODING) as f:
-                text = f.read()
-
-    # Extract hanzi from text (with duplicates)
-    hanzi_list = filter_text(text)
+    # Extract hanzi from HTML (with duplicates)
+    hanzi_list = filter_text(html)
     # Divide into groups (with duplicates)
     simplified, traditional, neutral, outliers = partition_hanzi(
         hsk_simp, hsk_trad, hanzi_list
@@ -95,10 +50,10 @@ def analyse_data(
     df: pd.DataFrame, hl: list, simplified: list, traditional: list, neutral: list
 ) -> tuple[str | pd.DataFrame]:
     """
-    Receives the output of process_data
-    Gets the character variant and statistical breakdowns
+    Receives output of process_data
+    Gets character variant and statistical breakdowns
         - number of unique characters and number of total characters
-            by grade, and cumulative figures for the entire content
+          by grade, and cumulative figures for the entire content
 
     Parameters
     ----------
@@ -107,6 +62,7 @@ def analyse_data(
 
     hl : list
         all hanzi in the entire content
+
     simplified : list
         simplified HSK hanzi in hl
